+++ src/components/Message/Message.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+import './Message.css';
+class Message extends Component {
+	render() {
+		return (
+			<div>
+				<span className="message">{this.props.text}</span>
+			</div>
+		);
+	}
+}
+
+export default Message;
